+++ ClothesShop/ViewModels/PdFilterVM.cs
@@ -19,7 +19,7 @@ namespace ClothesShop.ViewModels
 
         public double? PdCost { get; set; }
 
-        public double? PdStk { get; set; }
+        public int? PdStk { get; set; }
 
         public byte? PdtId { get; set; }
         public string? PdtName { get; set; }
